20240218 @Mookse - move avatar from class-extenders to single .mjs
