+++ inc/js/factory-class-extenders/class-extenders.mjs
@@ -1,15 +1,6 @@
 import { EventEmitter } from 'events'
 import { EvolutionAssistant } from '../agents/system/evolution-assistant.mjs'
 import {
-    mAI_openai,
-    mAssignEvolverListeners,
-    mBot,
-    mChat,
-    mFindBot,
-    mGetAssistant,
-    mGetChatCategory,
- } from './class-avatar-functions.mjs'
-import {
     mGetQuestions,
     mUpdateContribution,
 } from './class-contribution-functions.mjs'
@@ -25,284 +16,6 @@ import {
 import { _ } from 'ajv'
 import { parse } from 'path'
 //  function definitions to extend remarkable classes
-function extendClass_avatar(_originClass,_references) {
-    class Avatar extends _originClass {
-        #activeBotId // id of active bot in this.#bots; empty or undefined, then this
-        #activeChatCategory = mGetChatCategory()
-        #bots = []
-        #conversations = []
-        #emitter = new EventEmitter()
-        #evolver
-        #factory
-        #openai = _references?.openai
-        #proxyBeing = 'human'
-        constructor(_obj,_factory) {
-            super(_obj) //  should include contributions from db or from class schema
-            if(_obj.proxyBeing)
-                this.#proxyBeing = _obj.proxyBeing
-            this.#factory = _factory
-            this.#bots = _obj?.bots ?? [] // array of ids
-        }
-        async init(){
-            /* create evolver (exclude MyLife) */
-            // @todo: admin interface for modifying MyLife avatar and their bots
-            this.#bots = await this.factory.bots(this.id)
-            let _activeBot = this.avatarBot
-            if(!this.factory.isMyLife){
-                if(!_activeBot.id){ // create: but do not want to call setBot() to activate
-                    _activeBot = await mBot(this)
-                    this.#bots.unshift(_activeBot)
-                }
-                this.activeBotId = _activeBot.id
-                this.#evolver = new EvolutionAssistant(this)
-                mAssignEvolverListeners(this.#evolver, this)
-                /* init evolver */
-                await this.#evolver.init()
-            } else { // Q-specific
-                this.activeBotId = this.avatarBot.id
-                this.avatarBot.bot_id = (process.env.OPENAI_MAHT_GPT_OVERRIDE) ? process.env.OPENAI_MAHT_GPT_OVERRIDE : this.avatarBot.bot_id
-                let _conversation = await this.getConversation()
-                this.avatarBot.thread_id = _conversation.threadId
-            }
-            this.emit('avatar-init-end',this)
-            return this
-        }
-        /* public functions */
-        /**
-         * Get a bot.
-         * @public
-         * @param {string} _bot_id - The bot id.
-         * @returns {object} - The bot.
-         */
-        async bot(_bot_id){
-            return await this.factory.bot(_bot_id)
-        }
-        /**
-         * Processes and executes incoming chat request.
-         * @public
-         * @param {object} ctx - The context object.
-         * @returns {object} - The response(s) to the chat request.
-        */
-        async chatRequest(ctx){
-            if(!ctx?.state?.chatMessage)
-                throw new Error('No message provided in context')
-            const _chat = await mChat(
-                this,
-                ctx.state.chatMessage
-            )
-            const _activeAlerts = ctx.state.MemberSession.alerts()
-            if(_activeAlerts?.length){
-                _chat.alerts = ctx.state.MemberSession.alerts()
-            }
-            return _chat
-        }
-        /**
-         * Proxy for emitter.
-         * @public
-         * @param {string} _eventName - The event to emit.
-         * @param {any} - The event(s) to emit.
-         * @returns {void}
-         */
-        emit(_eventName, ...args){
-            this.#emitter.emit(_eventName, ...args)
-        }
-        async getAssistant(_dataservice){	//	openai `assistant` object
-            // @todo: move to modular function and refine instructions as other bots
-            if(!this.assistant?.id.length) {
-                this.assistant = await mAI_openai(this.#openai, {
-                    name: this.names?.[0]??this.name,
-                    model: process.env.OPENAI_MODEL_CORE_AVATAR,
-                    description: this.description,
-                    instructions: this.purpose,
-                })
-                //	save id to cosmos
-                _dataservice.patch(this.id, {
-                    assistant: {
-                        id: this.assistant.id
-                    ,	object: 'assistant'
-                    }
-                })
-            } else if(!this.assistant.name?.length){
-                this.assistant = await mGetAssistant(this.#openai, this.assistant.id)
-            }
-            return this.assistant
-        }
-        async getConversation(_thread_id){ // per bot
-            let _conversation = this.#conversations.find(_=>_.thread?.id===_thread_id)
-            if(!_thread_id || !_conversation){
-                _conversation = new (this.factory.conversation)({ mbr_id: this.mbr_id}, this.factory)
-                await _conversation.init(_thread_id)
-                this.#conversations.push(_conversation)
-            }
-            return _conversation
-        }
-        on(_eventName, listener){
-            this.#emitter.on(_eventName, listener)
-        }
-        /**
-         * Processes and executes incoming category set request.
-         * @public
-         * @param {string} _category - The category to set { category, contributionId, question }.
-         */
-        setActiveCategory(_category){
-            const _proposedCategory = mGetChatCategory(_category)
-            /* evolve contribution */
-            if(_proposedCategory?.category){ // no category, no contribution
-                this.#evolver.setContribution(this.#activeChatCategory, _proposedCategory)
-            }
-        }
-        /**
-         * Add or update bot, and identifies as activated, unless otherwise specified.
-         * @param {object} _bot - Bot-data to set.
-         */
-        async setBot(_bot, _activate=true){
-            _bot = await mBot(this, _bot) // returns bot object
-            /* add bot to avatar */
-            const _index = this.#bots.findIndex(bot => bot.id === _bot.id)
-            if(_index !== -1) // update
-                this.#bots[_index] = _bot
-            else // add
-                this.#bots.push(_bot)
-            /* activation */
-            if(_activate)
-                this.activeBotId = _bot.id
-            return _bot
-        }
-        async thread_id(){
-            // @todo: once avatar extends bot, keep this inside
-            if(!this.#conversations.length){
-                await this.getConversation()
-            }
-            return this.#conversations[0].threadId
-        }
-        /* getters/setters */
-        /**
-         * Get the active bot. If no active bot, return this as default chat engine.
-         * @public
-         * @returns {object} - The active bot.
-         */
-        get activeBot(){
-            return this.#bots.find(_bot=>_bot.id===this.activeBotId)
-        }
-        get aiId(){
-            return this.conversation[0]
-        }
-        /**
-         * Get the active bot id.
-         * @public
-         * @returns {string} - The active bot id.
-         */
-        get activeBotId(){
-            return this.#activeBotId
-        }
-        /**
-         * Set the active bot id. If not match found in bot list, then defaults back to this.id
-         * @public
-         * @param {string} _bot_id - The active bot id.
-         * @returns {void}
-         */
-        set activeBotId(_bot_id){ // default PA
-            if(!_bot_id?.length) _bot_id = this.avatarBot.id
-            this.#activeBotId = mFindBot(this, _bot_id)?.id??this.avatarBot.id
-        }
-        /**
-         * Returns provider for avatar intelligence.
-         * @public
-         * @returns {object} - The avatar intelligence provider, currently only openAI API GPT.
-         */
-        get ai(){
-            return this.#openai
-        }
-        get avatar(){
-            return this.inspect(true)
-        }
-        get avatarBot(){
-            return this.#bots.find(_bot=>_bot.type==='personal-avatar')
-        }
-        /**
-         * Get the "avatar's" being, or more precisely the name of the being (affiliated object) the evatar is emulating.
-         * Avatars are special case and are always avatars, so when we query them non-internally for system purposes (in which case we understand we need to go directly to factory.core.being) we display the underlying essence of the datacore; could put this in its own variable, but this seems protective _and_ gives an access point for alterations.
-         * @returns {string} The object being the avatar is emulating.
-        */
-        get being(){    //  
-            return this.#proxyBeing
-        }
-        get bots(){
-            return this.#bots
-        }
-        get category(){
-            return this.#activeChatCategory
-        }
-        set category(_category){
-            this.#activeChatCategory = _category
-        }
-        get contributions(){
-            return this.#evolver?.contributions
-        }
-        /**
-         * Set incoming contribution.
-         * @param {object} _contribution
-        */
-       set contribution(_contribution){
-           this.#evolver.contribution = _contribution
-        }
-        /**
-         * Get uninstantiated class definition for conversation. If getting a specific conversation, use .conversation(id).
-         * @returns {class} - class definition for conversation
-         */
-        get conversation(){
-            return this.factory.conversation
-        }
-        /**
-         * Get conversations. If getting a specific conversation, use .conversation(id).
-         * @returns {array} - The conversations.
-         */
-        get conversations(){
-            return this.#conversations
-        }
-        // todo: deprecate to available convenience public emit() function
-        get emitter(){  //  allows parent to squeeze out an emitter
-            return this.#emitter
-        }
-        // @todo: return factory privacy
-        get factory(){
-            return this.#factory
-        }
-        get globals(){
-            return this.factory.globals
-        }
-        get mbr_id(){
-            return this.factory.mbr_id
-        }
-        get mbr_id_id(){
-            return this.mbr_sysId
-        }
-        get mbr_name(){
-            return this.mbr_sysName
-        }
-        get mbr_sysId(){
-            return this.globals.sysId(this.mbr_id)
-        }
-        get mbr_sysName(){
-            return this.globals.sysName(this.mbr_id)
-        }
-        get memberFirstName(){
-            return this.memberName.split(' ')[0]
-        }
-        get memberName(){
-            return this.factory.memberName
-        }
-        /**
-         * Get uninstantiated class definition for message.
-         * @returns {class} - class definition for message
-         */
-        get message(){
-            return this.factory.message
-        }
-    }
-    console.log('Avatar class extended')
-    return Avatar
-}
 function extendClass_consent(_originClass,_references) {
     class Consent extends _originClass {
         constructor(_obj) {
@@ -566,7 +279,6 @@ function extendClass_message(_originClass,_references) {
 }
 /* exports */
 export {
-	extendClass_avatar,
 	extendClass_consent,
     extendClass_contribution,
     extendClass_conversation,
